Issue 558 (#559)
### What problem were solved in this pull request?

Issue Number: #558 

Problem:
As I mentioned in the issue, students cannot use the arrow keys to
browse the command history in the current cli. The existing code seems
to want to use `readline`(GNU Readline library), which may conflict with
the existing license. So I used replxx to improve this function in the
cli.

### What is changed and how it works?
#### Key changes:
- 3rd party dependency: replxx has been added as a git submodule
- `CMakeLists.txt` in `observer`
- related cpp source files
#### Basic logic:
1. The `CliCommunicator `class utilizes a `replxx::Replxx` instance to
manage the interaction.
2. Every 5s, history saved to `./.miniob.history`.

### Other information
1. This PR focuses **solely** on the `observer -P cli` functionality.
The **obclient** is **NOT** modified.
2. Default max_history_size = 1000.
3. Default config: when `current_time - previous_history_save_time > 5`,
history_save.