fix: semantic conflict of reset_data. (#583)
reset_data() only reset the count, but not the data. When processing the
string-like value, the previous data will be cached, and they may
interfere with the new value. So reset the data after the count in
reset_data().

### What problem were solved in this pull request?

Problem:

After calling reset_data() of Column with string type, call
append_value() with another Value(). As long as the attr_len_ >
Value.size(), the previous data will be retained, and they may interfere
with the newly one, as the reset_data() does not clear the actual data
inside.

For example:
id
ida
idb

After calling reset_data() and then call append_value() with actual
string "a", the column will be as follows:
ad(the char 'c' retains.)
(the following are invalid.)

### What is changed and how it works?

reset_data() will clear the inside data actually, not only reset the
count.